feat: add option to disable rounding
This solves the problem with verification of calculated gross income. The rounding introduced small, yet problematic differences between net income and verification net income. Thus, resulting gross income was a little bit off. A little bit that was noticeable.

However, when using the rounding off calculation, the results are not exactly like it would be when calculated by the accounting software. For our use, it is ok though. And there is still the option to calculate correct net income with all the roundings on.